fix broken logging and adapt xunit example tests to be a multi service test

+++ examples/src/example_readings_aspnet_web_app/Services/DevicesService.cs
@@ -65,7 +65,7 @@ public class DevicesService : IDevicesService
                     VALUES (@name)
                     RETURNING id";
                 cmd.Parameters.AddWithValue("@name", request.Name);
-                deviceId = (int)(await cmd.ExecuteScalarAsync()!)!;
+                deviceId = (int)(long)(await cmd.ExecuteScalarAsync()!)!;
             }
 
             await _workflowService.StartDeviceAsync(deviceId);
